--- src/components/Login/Login.scss
@@ -1,46 +1,0 @@
-.login {
-  display: flex;
-  flex: auto;
-  flex-direction: column;
-  justify-content: flex-start;
-
-  box-sizing: border-box;
-  width: 100%;
-  max-width: 396px;
-  height: 100%;
-  max-height: 553px;
-  margin: 0;
-  padding: 0;
-
-  &__title {
-    margin: 0 0 40px;
-    padding: 0;
-
-    font-size: 24px;
-    font-weight: 500;
-    line-height: 1.21; // 29px
-    text-align: left;
-    letter-spacing: 0;
-
-    @media screen and (width <= 350px) {
-      text-align: center;
-    }
-
-    @media screen and (width <= 320px) {
-      margin: 0 0 80px;
-    }
-  }
-
-  @media screen and (width <= 768px) {
-    max-height: 500px;
-  }
-
-  @media screen and (width <= 487px) {
-    max-width: 81%;
-  }
-
-  @media screen and (width <= 320px) {
-    max-width: 260px;
-    max-height: 636px;
-  }
-}
