+++ src/sass/auth.scss
@@ -1,4 +1,4 @@
-.register {
+.auth {
   display: flex;
   flex: auto;
   flex-direction: column;
